Split oversized test files into focused modules
test_settings.py (1312 lines) split into:
  - test_settings.py (706 lines): module and autorun processor tests
  - test_settings_api_routes.py (533 lines): Flask API route tests
  - test_settings_directories.py (104 lines): directory path settings

test_detectors.py (1544 lines) split into:
  - test_detectors.py (541 lines): sort, autorun, autodetect, progress
  - test_detector_find.py (787 lines): find-label and demo-origin
  - test_detector_export.py (247 lines): server export and labelset export

test_importers.py (1585 lines) split into:
  - test_importers.py (456 lines): base class, metadata, extract_archive
  - test_importer_loading.py (838 lines): load_dataset, content vectors/md5s
  - test_importer_symlinks.py (328 lines): symlink discovery and import

All test counts preserved (179 + 81 + 70). Test group markers updated.
