+++ tests/test_detector_find.py
@@ -1,3 +1,8 @@
+"""Detector find-label tests.
+
+Covers POST /api/find-label, demo-origin resolution, and stamp-demo-origin.
+"""
+
 import io
 import json
 
@@ -6,308 +11,6 @@ import pytest
 
 import app as app_module
 from helpers import train_detector_from_votes
-
-
-class TestDetectorSort:
-    def test_sort_with_valid_detector(self, client):
-        # Train a detector
-        app_module.good_votes.update({k: None for k in [1, 2]})
-        app_module.bad_votes.update({k: None for k in [3, 4]})
-        detector = train_detector_from_votes()
-
-        # Now use it to sort
-        resp = client.post("/api/detector-sort", json={"detector": detector})
-        assert resp.status_code == 200
-        data = resp.get_json()
-        assert "results" in data
-        assert "threshold" in data
-        assert len(data["results"]) == app_module.NUM_MEDIAS
-
-    def test_sort_results_sorted_descending(self, client):
-        app_module.good_votes.update({k: None for k in [1, 2]})
-        app_module.bad_votes.update({k: None for k in [3, 4]})
-        detector = train_detector_from_votes()
-
-        resp = client.post("/api/detector-sort", json={"detector": detector})
-        data = resp.get_json()
-        scores = [e["score"] for e in data["results"]]
-        assert scores == sorted(scores, reverse=True)
-
-    def test_sort_scores_in_valid_range(self, client):
-        app_module.good_votes.update({k: None for k in [1, 2]})
-        app_module.bad_votes.update({k: None for k in [3, 4]})
-        detector = train_detector_from_votes()
-
-        resp = client.post("/api/detector-sort", json={"detector": detector})
-        data = resp.get_json()
-        for entry in data["results"]:
-            assert 0.0 <= entry["score"] <= 1.0
-
-    def test_sort_missing_detector(self, client):
-        resp = client.post("/api/detector-sort", json={})
-        assert resp.status_code == 400
-
-    def test_sort_missing_weights(self, client):
-        resp = client.post("/api/detector-sort", json={"detector": {"threshold": 0.5}})
-        assert resp.status_code == 400
-
-    def test_detector_roundtrip(self, client):
-        """Export a detector and verify it produces reasonable scores."""
-        app_module.good_votes.update({k: None for k in [1, 2, 3]})
-        app_module.bad_votes.update({k: None for k in [18, 19, 20]})
-
-        # Export detector
-        detector = train_detector_from_votes()
-
-        # Use detector to sort
-        resp = client.post("/api/detector-sort", json={"detector": detector})
-        data = resp.get_json()
-        score_map = {e["id"]: e["score"] for e in data["results"]}
-
-        # Good medias should score higher than bad medias on average
-        avg_good = np.mean([score_map[i] for i in app_module.good_votes])
-        avg_bad = np.mean([score_map[i] for i in app_module.bad_votes])
-        assert avg_good > avg_bad
-
-
-class TestAutorunDetectors:
-    """Tests for the autorun-detectors management endpoints."""
-
-    def _export_detector(self, client):
-        """Helper: vote on some medias and train a valid detector payload."""
-        app_module.good_votes.update({k: None for k in [1, 2, 3]})
-        app_module.bad_votes.update({k: None for k in [18, 19, 20]})
-        return train_detector_from_votes()
-
-    def _post_autorun(self, client, name, detector):
-        return client.post(
-            "/api/autorun-detectors",
-            json={
-                "name": name,
-                "media_type": "audio",
-                "weights": detector["weights"],
-                "threshold": detector["threshold"],
-            },
-        )
-
-    # -- GET list --
-
-    def test_get_empty_list(self, client):
-        resp = client.get("/api/autorun-detectors")
-        assert resp.status_code == 200
-        assert resp.get_json()["detectors"] == []
-
-    def test_get_list_after_add(self, client):
-        det = self._export_detector(client)
-        self._post_autorun(client, "my-detector", det)
-
-        resp = client.get("/api/autorun-detectors")
-        data = resp.get_json()
-        assert len(data["detectors"]) == 1
-        d = data["detectors"][0]
-        assert d["name"] == "my-detector"
-        assert d["media_type"] == "audio"
-        assert "threshold" in d
-
-    # -- POST add --
-
-    def test_add_detector_returns_success(self, client):
-        det = self._export_detector(client)
-        resp = self._post_autorun(client, "test-det", det)
-        assert resp.status_code == 200
-        assert resp.get_json()["success"] is True
-
-    def test_add_missing_name_returns_400(self, client):
-        det = self._export_detector(client)
-        resp = client.post(
-            "/api/autorun-detectors",
-            json={"media_type": "audio", "weights": det["weights"]},
-        )
-        assert resp.status_code == 400
-
-    def test_add_missing_media_type_returns_400(self, client):
-        det = self._export_detector(client)
-        resp = client.post(
-            "/api/autorun-detectors",
-            json={"name": "test", "weights": det["weights"]},
-        )
-        assert resp.status_code == 400
-
-    def test_add_without_weights_creates_untrained(self, client):
-        resp = client.post(
-            "/api/autorun-detectors",
-            json={"name": "test", "media_type": "audio"},
-        )
-        assert resp.status_code == 200
-
-        detectors = client.get("/api/autorun-detectors").get_json()["detectors"]
-        det = [d for d in detectors if d["name"] == "test"][0]
-        assert det["weights"] is None
-        assert det["autodetect"] is False
-
-    def test_add_multiple_detectors(self, client):
-        det = self._export_detector(client)
-        self._post_autorun(client, "det-a", det)
-        app_module.good_votes.clear()
-        app_module.bad_votes.clear()
-        self._post_autorun(client, "det-b", det)
-
-        resp = client.get("/api/autorun-detectors")
-        names = {d["name"] for d in resp.get_json()["detectors"]}
-        assert names == {"det-a", "det-b"}
-
-    def test_add_overwrites_existing_name(self, client):
-        det = self._export_detector(client)
-        self._post_autorun(client, "dup", det)
-        self._post_autorun(client, "dup", det)
-
-        resp = client.get("/api/autorun-detectors")
-        assert len(resp.get_json()["detectors"]) == 1
-
-    # -- DELETE --
-
-    def test_delete_detector(self, client):
-        det = self._export_detector(client)
-        self._post_autorun(client, "to-delete", det)
-
-        resp = client.delete("/api/autorun-detectors/to-delete")
-        assert resp.status_code == 200
-        assert resp.get_json()["success"] is True
-
-        resp = client.get("/api/autorun-detectors")
-        assert resp.get_json()["detectors"] == []
-
-    def test_delete_nonexistent_returns_404(self, client):
-        resp = client.delete("/api/autorun-detectors/does-not-exist")
-        assert resp.status_code == 404
-
-    # -- RENAME --
-
-    def test_rename_detector(self, client):
-        det = self._export_detector(client)
-        self._post_autorun(client, "old-name", det)
-
-        resp = client.put(
-            "/api/autorun-detectors/old-name/rename",
-            json={"new_name": "new-name"},
-        )
-        assert resp.status_code == 200
-        assert resp.get_json()["new_name"] == "new-name"
-
-        names = [d["name"] for d in client.get("/api/autorun-detectors").get_json()["detectors"]]
-        assert "new-name" in names
-        assert "old-name" not in names
-
-    def test_rename_nonexistent_returns_400(self, client):
-        resp = client.put(
-            "/api/autorun-detectors/ghost/rename",
-            json={"new_name": "anything"},
-        )
-        assert resp.status_code == 400
-
-    def test_rename_to_existing_name_returns_400(self, client):
-        det = self._export_detector(client)
-        self._post_autorun(client, "det-a", det)
-        app_module.good_votes.clear()
-        app_module.bad_votes.clear()
-        self._post_autorun(client, "det-b", det)
-
-        resp = client.put(
-            "/api/autorun-detectors/det-a/rename",
-            json={"new_name": "det-b"},
-        )
-        assert resp.status_code == 400
-
-    def test_rename_missing_new_name_returns_400(self, client):
-        det = self._export_detector(client)
-        self._post_autorun(client, "some-det", det)
-
-        resp = client.put(
-            "/api/autorun-detectors/some-det/rename",
-            json={},
-        )
-        assert resp.status_code == 400
-
-    # -- import-pkl (detector JSON file) --
-
-    def test_import_pkl_from_detector_json(self, client):
-        det = self._export_detector(client)
-        json_bytes = json.dumps(det).encode("utf-8")
-        data = {
-            "file": (io.BytesIO(json_bytes), "detector.json"),
-            "name": "imported",
-        }
-        resp = client.post(
-            "/api/autorun-detectors/import-pkl",
-            data=data,
-            content_type="multipart/form-data",
-        )
-        assert resp.status_code == 200
-        result = resp.get_json()
-        assert result["success"] is True
-        assert result["name"] == "imported"
-
-    def test_import_pkl_uses_filename_stem_as_default_name(self, client):
-        det = self._export_detector(client)
-        json_bytes = json.dumps(det).encode("utf-8")
-        data = {"file": (io.BytesIO(json_bytes), "my_detector.json")}
-        resp = client.post(
-            "/api/autorun-detectors/import-pkl",
-            data=data,
-            content_type="multipart/form-data",
-        )
-        assert resp.status_code == 200
-        assert resp.get_json()["name"] == "my_detector"
-
-    def test_import_pkl_preserves_media_type_from_file(self, client):
-        det = self._export_detector(client)
-        # Embed explicit media_type in the "file" payload
-        det["media_type"] = "image"
-        json_bytes = json.dumps(det).encode("utf-8")
-        data = {
-            "file": (io.BytesIO(json_bytes), "image_detector.json"),
-            "name": "img-det",
-        }
-        resp = client.post(
-            "/api/autorun-detectors/import-pkl",
-            data=data,
-            content_type="multipart/form-data",
-        )
-        assert resp.status_code == 200
-        assert resp.get_json()["media_type"] == "image"
-
-    def test_import_pkl_no_file_returns_400(self, client):
-        resp = client.post("/api/autorun-detectors/import-pkl", data={})
-        assert resp.status_code == 400
-
-    def test_import_pkl_invalid_format_returns_400(self, client):
-        data = {"file": (io.BytesIO(b'{"not_a_detector": true}'), "bad.json")}
-        resp = client.post(
-            "/api/autorun-detectors/import-pkl",
-            data=data,
-            content_type="multipart/form-data",
-        )
-        assert resp.status_code == 400
-
-    # -- Detector data is stored correctly --
-
-    def test_stored_detector_has_correct_fields(self, client):
-        det = self._export_detector(client)
-        self._post_autorun(client, "field-check", det)
-
-        from vtsearch.utils.state import autorun_detectors
-
-        assert "field-check" in autorun_detectors
-        stored = autorun_detectors["field-check"]
-        assert stored["name"] == "field-check"
-        assert stored["media_type"] == "audio"
-        assert "weights" in stored
-        assert "threshold" in stored
-        assert "created_at" in stored
-        assert "good_origins" in stored
-        assert "bad_origins" in stored
-        assert "inclusion" in stored
-
 
 class TestFindLabel:
     """Tests for POST /api/find-label — score all medias and apply Good/Bad labels."""
@@ -438,7 +141,7 @@ class TestFindLabel:
         Previously returned 400 because no pre-trained weights existed.
         """
         from vtsearch.models.registry import register_model, reset_for_tests
-        from vtsearch.models.trainable_model_store import _write_model
+        from vtsearch.routes.trainable_models import _write_model
         from vtsearch.datasets.labelset import LabelSet
         from vtsearch.utils import bad_votes, good_votes, snapshot_medias
 
@@ -505,7 +208,7 @@ class TestFindLabel:
         from unittest.mock import patch
 
         from vtsearch.models.registry import register_model, reset_for_tests
-        from vtsearch.models.trainable_model_store import _write_model
+        from vtsearch.routes.trainable_models import _write_model
         from vtsearch.settings import get_trainable_models_dir, set_trainable_models_dir
         from vtsearch.utils import medias
 
@@ -612,7 +315,7 @@ class TestFindLabel:
     def test_find_label_cross_dataset_error_includes_diagnostics(self, client, tmp_path):
         """When origin resolution fails, the error response should include diagnostics."""
         from vtsearch.models.registry import register_model, reset_for_tests
-        from vtsearch.models.trainable_model_store import _write_model
+        from vtsearch.routes.trainable_models import _write_model
         from vtsearch.settings import get_trainable_models_dir, set_trainable_models_dir
         from vtsearch.utils import medias
 
@@ -698,8 +401,7 @@ class TestFindLabel:
         """
         from vtsearch.datasets.labelset import LabelSet
         from vtsearch.models.registry import add_loaded_model_id, register_model, reset_for_tests
-        from vtsearch.models.label_sync import sync_labels_to_loaded_model
-        from vtsearch.models.trainable_model_store import _read_model, _write_model
+        from vtsearch.routes.trainable_models import _read_model, _write_model, sync_labels_to_loaded_model
         from vtsearch.settings import get_trainable_models_dir, set_trainable_models_dir
         from vtsearch.utils import bad_votes, good_votes, set_thread_detector_context, snapshot_medias
         from vtsearch.utils.state_core import DetectorContext, register_detector_context
@@ -874,7 +576,7 @@ class TestFindLabelDemoOrigin:
         from unittest.mock import patch
 
         from vtsearch.models.registry import register_model, reset_for_tests
-        from vtsearch.models.trainable_model_store import _write_model
+        from vtsearch.routes.trainable_models import _write_model
         from vtsearch.settings import get_trainable_models_dir, set_trainable_models_dir
         from vtsearch.utils import medias
 
@@ -983,7 +685,7 @@ class TestFindLabelDemoOrigin:
         Simulates old pickles that stored demo origins without the dataset name.
         """
         from vtsearch.models.registry import register_model, reset_for_tests
-        from vtsearch.models.trainable_model_store import _write_model
+        from vtsearch.routes.trainable_models import _write_model
         from vtsearch.settings import get_trainable_models_dir, set_trainable_models_dir
         from vtsearch.utils import medias
 
@@ -1080,235 +782,5 @@ class TestStampDemoOrigin:
         # Mutating one should not affect the other
         medias[1]["origin"]["params"]["extra"] = "test"
         assert "extra" not in medias[2]["origin"]["params"]
-
-
-class TestDetectorSortProgress:
-    """Tests for progress reporting during POST /api/detector-sort."""
-
-    def test_detector_sort_reports_progress(self, client):
-        """detector-sort should update find_progress while scoring."""
-        from vtsearch.utils.progress import find_progress
-
-        # Train a detector
-        app_module.good_votes.update({k: None for k in [1, 2, 3]})
-        app_module.bad_votes.update({k: None for k in [18, 19, 20]})
-        detector = train_detector_from_votes()
-        app_module.good_votes.clear()
-        app_module.bad_votes.clear()
-
-        # Capture progress snapshots
-        snapshots = []
-        original_update = find_progress.update
-
-        def capturing_update(*args, **kwargs):
-            original_update(*args, **kwargs)
-            snapshots.append(find_progress.get())
-
-        find_progress.update = capturing_update
-        try:
-            resp = client.post("/api/detector-sort", json={"detector": detector})
-        finally:
-            find_progress.update = original_update
-
-        assert resp.status_code == 200
-
-        # Should have scoring progress snapshots
-        scoring = [s for s in snapshots if s.get("step") == 1 and s["status"] == "running"]
-        assert len(scoring) >= 1
-        assert "Scoring" in scoring[0]["message"]
-        assert scoring[0]["total"] == app_module.NUM_MEDIAS
-
-        # Should end with complete progress
-        final_running = [s for s in snapshots if s["current"] == app_module.NUM_MEDIAS and s["status"] == "running"]
-        assert len(final_running) >= 1
-
-        # Final state should be idle
-        final = find_progress.get()
-        assert final["status"] == "idle"
-
-
-class TestAutoDetect:
-    """Tests for POST /api/auto-detect."""
-
-    def _add_audio_detector(self, client, name="test-detector"):
-        """Helper: create and save an audio detector with autodetect enabled."""
-        app_module.good_votes.update({k: None for k in [1, 2, 3]})
-        app_module.bad_votes.update({k: None for k in [18, 19, 20]})
-        detector = train_detector_from_votes()
-
-        save_resp = client.post(
-            "/api/autorun-detectors",
-            json={
-                "name": name,
-                "media_type": "audio",
-                "weights": detector["weights"],
-                "threshold": detector["threshold"],
-                "autodetect": True,
-            },
-        )
-        assert save_resp.status_code == 200
-        app_module.good_votes.clear()
-        app_module.bad_votes.clear()
-
-    # -- no matching detectors --
-
-    def test_no_autorun_detectors_returns_400(self, client):
-        resp = client.post("/api/auto-detect")
-        assert resp.status_code == 400
-        assert "No autorun detectors" in resp.get_json()["error"]
-
-    def test_no_matching_media_type_returns_400(self, client):
-        """A detector for a different media type should not match audio medias."""
-        app_module.good_votes.update({k: None for k in [1, 2, 3]})
-        app_module.bad_votes.update({k: None for k in [18, 19, 20]})
-        detector = train_detector_from_votes()
-        app_module.good_votes.clear()
-        app_module.bad_votes.clear()
-
-        # Save as "image" type with autodetect — medias are audio, so it won't match
-        client.post(
-            "/api/autorun-detectors",
-            json={
-                "name": "image-detector",
-                "media_type": "image",
-                "weights": detector["weights"],
-                "threshold": detector["threshold"],
-                "autodetect": True,
-            },
-        )
-        resp = client.post("/api/auto-detect")
-        assert resp.status_code == 400
-
-    # -- basic success --
-
-    def test_returns_200_with_matching_detector(self, client):
-        self._add_audio_detector(client)
-        resp = client.post("/api/auto-detect")
-        assert resp.status_code == 200
-
-    def test_response_has_required_top_level_fields(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        assert "media_type" in data
-        assert "detectors_run" in data
-        assert "results" in data
-
-    def test_media_type_matches_clips(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        assert data["media_type"] == "audio"
-
-    def test_detectors_run_count(self, client):
-        self._add_audio_detector(client, name="det-1")
-        self._add_audio_detector(client, name="det-2")
-        data = client.post("/api/auto-detect").get_json()
-        assert data["detectors_run"] == 2
-
-    # -- per-detector result structure --
-
-    def test_each_result_has_required_fields(self, client):
-        self._add_audio_detector(client, name="struct-check")
-        data = client.post("/api/auto-detect").get_json()
-        result = data["results"]["struct-check"]
-        assert "detector_name" in result
-        assert "threshold" in result
-        assert "total_hits" in result
-        assert "hits" in result
-
-    def test_detector_name_matches_key(self, client):
-        self._add_audio_detector(client, name="named-detector")
-        data = client.post("/api/auto-detect").get_json()
-        result = data["results"]["named-detector"]
-        assert result["detector_name"] == "named-detector"
-
-    def test_total_hits_matches_hits_length(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        for result in data["results"].values():
-            assert result["total_hits"] == len(result["hits"])
-
-    # -- hit data safety --
-
-    def test_hits_do_not_contain_embeddings(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        for result in data["results"].values():
-            for hit in result["hits"]:
-                assert "embedding" not in hit
-
-    def test_hits_do_not_contain_media_bytes(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        for result in data["results"].values():
-            for hit in result["hits"]:
-                assert "media_bytes" not in hit
-
-    def test_hits_contain_score(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        for result in data["results"].values():
-            for hit in result["hits"]:
-                assert "score" in hit
-                assert 0.0 <= hit["score"] <= 1.0
-
-    def test_hits_sorted_descending_by_score(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        for result in data["results"].values():
-            scores = [h["score"] for h in result["hits"]]
-            assert scores == sorted(scores, reverse=True)
-
-    # -- threshold correctness --
-
-    def test_all_hits_score_at_or_above_threshold(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        for result in data["results"].values():
-            threshold = result["threshold"]
-            for hit in result["hits"]:
-                assert hit["score"] >= threshold - 1e-6  # float tolerance
-
-    # -- negative_hits --
-
-    def test_each_result_has_negative_hits(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        for result in data["results"].values():
-            assert "negative_hits" in result
-            assert isinstance(result["negative_hits"], list)
-
-    def test_negative_hits_score_below_threshold(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        for result in data["results"].values():
-            threshold = result["threshold"]
-            for hit in result["negative_hits"]:
-                assert hit["score"] < threshold + 1e-6
-
-    def test_negative_hits_do_not_contain_embeddings(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        for result in data["results"].values():
-            for hit in result["negative_hits"]:
-                assert "embedding" not in hit
-                assert "media_bytes" not in hit
-
-    def test_hits_and_negative_hits_cover_all_clips(self, client):
-        """Positive + negative hits should cover every media in the dataset."""
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        from vtsearch.utils import medias
-
-        total_clips = len(medias)
-        for result in data["results"].values():
-            total_returned = len(result["hits"]) + len(result["negative_hits"])
-            assert total_returned == total_clips
-
-    def test_negative_hits_sorted_descending_by_score(self, client):
-        self._add_audio_detector(client)
-        data = client.post("/api/auto-detect").get_json()
-        for result in data["results"].values():
-            scores = [h["score"] for h in result["negative_hits"]]
-            assert scores == sorted(scores, reverse=True)
